Move to a src/ dir with a setup.py.
