+++ .gitignore
@@ -1,1 +1,4 @@
 *.pyc
+*.egg-info
+build
+dist
