+++ src/fulgurate/argopen.py
